refactor: using api route
